// front-coin-wallet/src/app/services/com.b2beyond.api.webwallet-service-b2bcoin/model/walletKeys.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface WalletKeys {
    privateSpendKey?: string;
    privateViewKey?: string;
    publicViewKey?: string;
}