// front-coin-wallet/src/app/services/com.b2beyond.api.webwallet-service-b2bcoin/model/walletCoin.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface WalletCoin {
    active?: boolean;
    coinBaseUrl?: string;
    convertAmount?: number;
    fee?: number;
    feeActiveForExchangeMembers?: boolean;
    fractionDigits?: number;
    logo?: string;
    minTransactionsForNoFee?: number;
    monero?: boolean;
    name?: string;
    primaryCoin?: boolean;
    rpcPortDaemon?: number;
    rpcPortWallet?: number;
    walletBaseUrl?: string;
}